Migrations issues resolved

+++ database/migrations/2025_09_04_102918_remove_email_phone_description_from_departments_table.php
@@ -13,9 +13,11 @@ class RemoveEmailPhoneDescriptionFromDepartmentsTable extends Migration
      */
     public function up()
     {
-        Schema::table('departments', function (Blueprint $table) {
+        if (Schema::hasTable('departments') && !Schema::hasColumn('departments', 'email')) {
+            Schema::table('departments', function (Blueprint $table) {
             $table->dropColumn(['email', 'phone', 'description']);
         });
+        }
     }
 
     /**
